Make ornode dev server reload when dependencies change
Tried to solve the same problem for gui. Currently you have to re-save
the vite.config.ts file in order for dev server to rebuild with changed
dependencies. Had to add force flag and optimizeDeps.force: true to
vite.config.ts to make this work. This problem is bugging a lot of vite
users: https://github.com/vitejs/vite/issues/8619 .

Also refactored clientToNodeTransformer to not depend on z.instanceof,
because it is not reliable...