Implement proposals to execute delegatecalls
This makes Orec extensible.

For example if you want to make sure proposals A and B are executed in
order, you can deploy another contract that has a function that executes
proposals in expected order. Then Orec proposal should be passed to
delegatecall that new contract and that's it! Whatever the
delegatecalled contract will do / call, it will do so on behalf of Orec
(meaning any owner checks for Orec as the owner will pass).